render: sub-pixel anginc for ray-plane density control
anginc (the angular ray-step divisor, j_count = range / anginc) was an
i32 floored at 1, so the scan fan could only be coarsened. Both consumers
already read it as f32, so retype OpticastSettings/ScanContext.anginc to
f32. anginc < 1 now supersamples the angular fan (more ray planes), which
hides the thin-geometry silhouette holes the grouscan marcher drops on
isolated voxels. This is a density knob, not a fix: the underlying CPU
marcher miss is still there, just sampled over.

anginc = 1.0 is byte-identical to the old integer 1, so all goldens are
unchanged. Size the radar/angstart scratch for the inflated ray count
(CpuBackend pool and the per-frame path) so hrend stays in bounds when
anginc < 1.

+++ crates/roxlap-core/src/opticast.rs
@@ -58,7 +58,7 @@ pub struct OpticastSettings {
     pub hx: f32,
     pub hy: f32,
     pub hz: f32,
-    pub anginc: i32,
+    pub anginc: f32,
     pub mip_levels: u32,
     pub mip_scan_dist: i32,
     pub max_scan_dist: i32,
@@ -85,7 +85,7 @@ impl OpticastSettings {
             hx: half_w,
             hy: half_h,
             hz: half_w,
-            anginc: 1,
+            anginc: 1.0,
             mip_levels: 1,
             mip_scan_dist: 4,
             max_scan_dist: 1024,
@@ -579,7 +579,7 @@ mod tests {
         assert!((s.hx - 320.0).abs() < f32::EPSILON);
         assert!((s.hy - 240.0).abs() < f32::EPSILON);
         assert!((s.hz - 320.0).abs() < f32::EPSILON);
-        assert_eq!(s.anginc, 1);
+        assert_eq!(s.anginc, 1.0);
         assert_eq!(s.mip_levels, 1);
         assert_eq!(s.max_scan_dist, 1024);
     }
